MTU: Don't bounce the port when changing the MTU
When changing the MTU of a port to jumbo size or from jumbo
size to normal then for certain drivers there is no need to bounce
the port.  This is also the case on hardware switch platforms. The
MTU change should be be hitless when possible.

VRVDR-51148

(cherry picked from commit ed6dcba73c675248e54db3a6a806b95c3f86268f)