MTU: Register QoS's intereset in MTU change
MTU changes can affect the QoS shaping functionality due to the changes
in burst size.  As a result when MTU changes are notified to QoS, for
the software implementation, QoS scheduler is stopped and then started
with the revised burst sizes.  There is no change required for the HW
implmentation of QoS as it is able to apply the changes dynamically.

VRVDR-51148