cgnat: Increase maximum configurable max-dest-per-session to 128
VRVDR-50470.  Now that max-dest-per-session is a 'hard' maximum (in that
further flows are blocked), change the max configurable value from 64 to
128.  LEave the default max at 64.   We want to set a sensible default
max, but allow for a larger max to be configured.  An unecessarily large
default value is inefficient given that there may be a hash table of this
size per 3-tuple session, and there may be up to 30 million 3-tuple
sessions.