npf: Set custom timeout in dataplane session after session is created
VRVDR-51596.  npf has a configurable custom timeout. This was being set
from se_create by calling an npf function.  That npf function was
re-caching the packet in order to lookup the NPF_RS_CUSTOM_TIMEOUT
ruleset.  The packet was already cached before se_create was called, so
move the setting of custom timeout to where the same cache can be used.