npf: Add option to cache pkt without updating the cache grouper data
VRVDR-51596.  The grouper data in the packet cache is only needs to be
updated if the cache is subsequently used in a ruleset lookup.
Rename npf_cache_all_at to _npf_cache_all_at, and add a parameter to it
such that the grouper update is optional.  Add a wrapper function which
sets the grouper update parameter to false.