dpdk_eth_if: don't dump ports that have been unplugged
Once a port is marked as unplugged it will shortly be going away.
Don't dump ports that are marked as unplugged. This is necessary
as the dpdk dev data can be freed as part of this, and it is not
done under rcu control.

VRVDR-51100

(cherry picked from commit ee812da695b26b8f758b9c37bbf223086150cccd)