if: in ifnet_byethname skip over unplugged interfaces
Once a port is marked as unplugged it will shortly be going away.
Don't check prts that are marked as unplugged as the check is using
dpdk dev->data->name which can be be freed/cleared as part of
unplugging, and this is not done under rcu control.

VRVDR-51100

(cherry picked from commit c1ab919ae502fc6ea88f89b643bb0983526fa5c7)