hotplug: mark the interface as unplugged at the start of processing
Mark the interface as being unplugged before we start to process
the unplugging. Without this there is a small window where a thread
can check for unplugged, find it is not, and proceed and the access
dev->data->name.  If the hotplug thread then causes the interace to
be fully removed before the other thread has acessed dev->data->name
we can access invalid memory.  By moving the setting of unplugged
to the start of the removal, we get a call to synchronize_rcu
before the cleanup is complete and that means that the other threads
will not be able to access the dev->data->name once it is no longer
valid.

VRVDR-51100

(cherry picked from commit c85408af93c4d2c56c2a9186164267c38d64b28c)