if: clean up life-cycle of DPDK ethernet interface objects
Previously, DPDK ethernet interface objects were created at startup,
as opposed to when the netlink NEWLINK message arrives, meaning that
the life-cycle is messy and the interfaces have to be renamed and have
to be modified into being complete interfaces when the NEWLINK message
arrives, including being renamed (since the name wasn't previously
known) and ifindex being set. Furthermore, features that deal with
ifindices now need special notifications for when this information
becomes valid/invalid outside of the lifecycle of the interface, which
makes things more complex than they need to be.

There is no need for the interface objects of DPDK ethernet interfaces
to be created in advance of the netlink NEWLINK messages arriving, now
that there is no race with the if name to port mapping known compared
to when the NEWLINK message arrives. Therefore, simplify the
life-cycle of DPDK ethernet interfaces and make them consistent with
all other interfaces by creating them in response the the netlink
NEWLINK notification, rather than when discovered. The discovery
process still kickstarts the registration with the controller, which
will then drive the shadow interface to be created and the netlink
NEWLINK notification arriving. In addition, now delete the interface
objects in response to netlink DELLINK messages, with the
re-registration with the controller triggering re-creation of the
interface.

Note that this will result in a behaviour change when DPDK ethernet
ports have been discovered but no NEWLINK message has yet arrived,
whereby they won't be output by ifconfig. However, due to their names
not really being user identifiable and this either being the sign of a
race, or being due to a bug (with ifconfig results being limited to
not useful in debugging further) then this behaviour isn't maintained.

Additionally, fix up the unit tests to expect the new behaviour with
the life-cycle of interfaces, particularly with the switch port
interfaces in the tests for which the creation of these is delayed
until the tests that make use of them.

Closes: VRVDR-51844