devinfo: change if_port_info to not require an ifp
Change if_port_info to not require an ifp so that it can be used prior
to the creation of an interface object, when only the portid is known.

Rename the function to reflect its purpose as a DPDK-ethernet specific
function and move the source file to the src/if/dpdk-eth directory to
match.