ut: allow user to specify platform conf file
It is useful to allow users to provide their own fal plugin file,
and that is specified in the platform.conf file.  Allow the users to
be provide the location of hte platform.conf file.

Move the declaration of dp_args further down in the func so that it
uses the value of dp_test_platform_file after the config has been
parsed.

VRVDR-52113

(cherry picked from commit e7d794fded3bd598a470f8467b656d329b818e45)