qos: update fal global map when resource group changes
When a DSCP resource group changes, qos gets notified and
updates all policies using that resource group by removing them and
then reinstalling them. In the case where we are using a "global"
ingress map obj, it applies to all policies and needs to be changed.
The easiest way to do this is to delete it if the policies were removed
and it will be recreated by the first policy reinstall.

VRVDR-51940