shadow: remove superfluous interface netlink state management
The interface netlink state management is now done by the controller,
so remove the superfluous management of the netlink state in the shadow
code. Performing the creation/deletion of kernel interfaces in the
controller has the benefit of being able to purge stale interfaces
from the system after a dataplane restart where the configuration may
have changed in the intervening period, or otherwise the interfaces
aren't created again. It also means that it removes an obstacle to
cleaning up the ifp life-cycle in the dataplane.

VRVDR-51844