mstp: defer creation of STP object until after bridge created in FAL
Only create the FAL STP object referring to the bridge ifindex after
the FAL has been notified about the creation of the bridge interface
with that ifindex by making use of the IF_FEAT_MODE_EVENT_L2_CREATED
notification, and for symmetry and deletion in the correct order make
use of IF_FEAT_MODE_EVENT_L2_DELETED for triggering the deletion of
the FAL STP object prior to the bridge FAL L2 port being deleted.

Move the code into mstp.c to be common with other STP-object related
code.

Fixes: VRVDR-52083