route6: update dependent routes when interface FAL L3 state changes
The FAL L3 interface state is depended upon in
next_hop_to_packet_action, with the lack of a FAL router interface
object resulting in a TRAP packet action being programmed in the
route. Therefore, when the FAL L3 state changes for an interface
(resulting in the FAL router interface object being created or
deleted) then in order for the FAL plugin to react properly then it is
necessary to walk to dependent objects and update them, which in this
case is next-hop-groups and then routes.

Fixes: VRVDR-50303