if: fix cleanup of DPDK ethernet interfaces
Following commit 376a403ffc2dc0ce ("if: clean up life-cycle of DPDK
ethernet interface objects") it is expected that ifps will be created
when a RTM_NEWLINK message arrives for a dataplane interface, and it
is not expected that will already be an existing interface with the
same name. However, the cleanup code wasn't updated to delete the ifps
when the connection to the controller is reset and this breaks the
subsequent reconnect.

Therefore, fix if_cleanup to now delete all ifps regardless of their
type since now all interface types are treated equally.

Fixes: VRVDR-52145