ut: add vars that tests can use to pass state to the fal
It is useful for a test to be able to provide state to the fal
outwith the normal message passing.  Add an int and a ptr  that
can be set and read by the tests and the fal.

VRVDR-52113

(cherry picked from commit 633b2a7d50e07ecbf5d59f141e32b52a6897b908)