shadow: remove the shadow port handler when an interface is freed
After commit 230f1bd0f97c ("if: fix cleanup of DPDK ethernet
interfaces") there can exist a period of time where there is no ifp
backing a shadow port because the shadow port handler is only removed
when an interface is hot-unplugged.

Fix this by ensuring that in both the case when the interface is freed
and when the interface is hot-unplugged (but not both) that the shadow
port handler is removed so that no packets can be received on the
shadow port when there is no ifp backing it. The shadow port handler
will be added during ifp alloc if the interface happens to be created
again (such as after reconnecting to the controller).

Fixes: VRVDR-52193