config: factor out and simplify PCI address parsing for backplane ports
Simplify the parsing of the backplane port values by splitting the PCI
address parsing out from the backplane port name parsing. This does
mean an extra memory allocation, but this is at startup so transient
usage isn't important for either fragementation or performance
reasons.

Factor out the PCI address parsing into a separate function so that it
can be reused.

VRVDR-52175