config: parse management_port platform.conf attribute
It is useful to be able to determine on a platform whether a given
port is a designated management port or not.

So Parse the management_port<n> key in the dataplane section of
platform.conf and parse the key into a PCI address and add this to a
list.

VRVDR-52175