dpdk-eth: add management port attribute to interface information
Add a field indicating whether or not a DPDK ethernet interface is a
management one to the json information dumped as part of ifconfig,
making use of the information provided by the platform.

This information can then be used by the rest of the system for how to
treat the interface.

VRVDR-52175