sanitizer: set ASAN_OPTIONS when sanitizer is used
When the dataplane is build with the sanitizer the package produces
an invalid protocol file as the dependencies are not processed correctly.
This is because the ASAN_OPTIONS env var is not set when using the
dataplane to generate the deps. Set the ASAN_OPTIONS as part of the
call of build/dataplane.