main: handle sparse port configurations
Until this point, we have always had a compact list of ports
at start up.  The number of available ports would reflect the
highest port id in the system. However, with HyperV and
accelerated ports some ports may be claimed by another driver
before the dataplane has a chance to examine them. So the
number of available ports may not reflect the number of
ports in the system.

To handle this, manually iterate the list of available ports so
we can find the highest port and correctly fill in the enabled
port mask. Where necessary, skip these ports that aren't owned
by the dataplane.

Bugfix: VRVDR-49805