storm_ctl: fix write to heap after free when deleting an instance
A profile can be deleted prior to an instance due to asynchronicities
in the configuration handling, and if this occurs it can cause a write
to the scp_instance_list list head of the profile after the memory has
already been freed.

Fix this by keeping the profile around if there are instances that
refer to it, even if the profile is otherwise empty. This already
happens in the case that an instance message is received out of order,
referring to a profile that hasn't yet been created.

Fixes: VRVDR-52115
(cherry picked from commit 623e0464bccf80e0d227f958e05c58d03f28ec39)