netvsc: set more reasonable queue lengths
In HyperV there may be a "data" VF that is used to accelerate
performance. Typically, this is a mlx5 device. The transmit
and receive routines in the mlx5 driver require the queue lengths
to be a power of 2.

The current values in netvsc have no particular history, they just
seem to work. However, they cause the mlx5 driver to corrupt packets.
For now, pick some new values that will work for both drivers.

Bugfix: VRVDR-49805