controller: extend timeout for adding a port
Adding a port can take a little time for various reasons, e.g. due to
the work in the rest of the system triggered by the creation of the
port, or other pending state to be processed in the replay from the
controller. Therefore, it can take more than the 10 seconds to process
the response to the INIPORT/ADDPORT messages in the dataplane.

So to solve this extend the timeout to 1 minute which should give a
very wide margin, whilst still detecting errors of ports not being
added, should that happen for some reason.

Fixes: VRVDR-52458