netvsc: increase driver limits
The netvsc driver would fail to start if more than one queue was in
use. This was fixed in DPDK:

    net/netvsc: do not configure RSS if disabled

    This fixes the problem where driver would not start if only
    have a single Rx queue and multiple Txq. In that case, RSS
    should stay disabled.

and

    net/netvsc: do RSS across Rx queue only

    If number of tx queues is greater than the number of rx queues;
    the driver ends up allocating more channels than rx queues.
    The problem is that the RSS indirection table is programmed such
    that some packets will end up on a channel that would never be
    polled. The fix is to limit the RSS indirection table by number
    of rx queues not channels.

Prior to this point, netvsc was always a software PMD and using
more than one RX was unlikely to yield significant performance
gains. Now that we have SRIOV acceleration, the underlying
VF can likely take advantage of increased limits.

Bugfix: VRVDR-52360