protobuf: add Path message fields for recursive labels
Add a field for via_mpls_label which indicates that the path can
instead be resolved recursively via an MPLS label, instead of directly
by nexthop interface with a label stack, and although not used by the
dataplane can be used by other parts of the system talking in terms of
Paths. Along with this, add a n_non_recursive_labels field that
indicates how many of the labels in the label stack should be removed
if recursing via the local label instead.

VRVDR-52445