fal_plugin: add attributes for MPLS TTL mode
Add attributes for modelling changes to MPLS TTL modes. The attributes
are only provided at the switch level, rather than at a per-LSP level
since it seems unlikely that an operator would want to mix and match
between modes in a network. Similarly, no distinction is made between
ingress or egress modes as this is unlikely to have a use in a
well-functioning network.

Closes: VRVDR-52435