ptp: check switch nexthop interface is reachable
If a PTP packets needs to be routed (instead of directly switched)
we send it to the logical l3 interface on the bridge for routing.
If this interface happens to be "down", it's likely that you don't
have any routes for that interface either.

The solution is to create a master on each PTP port. Unfortunately,
the vendor stack only indexes masters by the IP address and not (IP
address, PTP port) tuple. In the FAL we already track all the masters
and (re)program the master at the top of the list.  However, we need to
remove any master that will likely fail to be to able contact its peer.

Bugfix: VRVDR-52447