event: add a public event notifier for vrf create/delete events
There is a private set of events that users can be notified about
but nothing public. Some of the events in the private set should
ideally be reworked before being made public, so create a minimal
set of public events that users can register for.

Use similar APIs, and internally have the public APIs use the internal
ones, by creating the internal set of events and copying the
public ones into the private version. Force the public and private
enum definitions to stay in sync with static asserts.

VRVDR-51860