qos: don't enable dscp mapping if designation in use
There is a default assumption that if a local priority queue
is in use then the profile will use DSCP mapping (as local
priority queue uses DSCP to classify packets as being of
sufficient priority). However, a local priority queue can
now be used with a designation mapping, so check for this
before applying the default.

VRVDR-52255