ippf: fix issue of corrupted first byte of IPv6 address to match
If the mask length is not on a byte boundary (i.e. is not a multiple of
8), in the "mixed byte" code the mask pointer get incremented, but blen
is not decremented. Therefore the following code to clear the bytes,
clears one byte too many. In the data structure, after the mask array,
is the array which contains the IPv6 address, so this results in the
first byte of the IP address being cleared.

This has been fixed by changing the code to use existing functions
to convert from a prefix to a bit mask - prefixlen_to_mask() for IPv4
and in6_prefixlen2mask() for IPv6.

VRVDR-52172