dpdk-eth: avoid duplicate link up/link down logs
Avoid duplicate link up/link down logs either caused by an interrupt
linkwatch update triggering which duplicates the state immediately at
the time of starting the port, or due to spurious interrupt triggers
later. Whilst we could take the position that the latter are DPDK PMD
issues and should be fixed there, it is easy to address both issues
with the same check on the previous port state.

So check the previous state of the port prior to notifying an update
as to the link state, but make sure to issue a log when the port is
started to give the operator a definitive notification about its state.

Fixes: VRVDR-52606