gre: don't double free mbuf if using a gre tunnel to ourself
The code in gre for a tunnel to ourself can cause a double free
if the incoming packet is fully processed.  In this case
where ip_gre[6]_tunnel_in() returns 0 we should return without
falling through to the drop case. If it returns non 0 then for
some reason it could not be processed and so needs to be
dropped.

VRVDR-50775