npf: Add a flag to the session to denote that is is being used as a fw session
VRVDR-52325.  This complements the bits that indicate snat, dnat, etc.
These flags are used for op-mode commands as a convenience to allow
filtering etc without querying the feature (npf).   A json bitmap
containing these values is also returned to the ctrl plane.  They are also
used for the session walk and session watch mechanisms.