flow_cache: don't init the flow cache multiple times per core
In the single processor case (the one the UTs use when running)
the forwarding_core is also core, 0, and so the lcore_events_init
func can be called as part of the main setup (for control on
core 0) and then again as part of making core 0 a forwadring
core.

Check if the flow cache is already initialised before creating new
structures.

VRVDR-52631