flow_cache: initialise the flow_cache_hash_key to 0
If we use the hash from the h_key that we have created then we
need to make sure that it has been initialised as the hash function
checks the full set of bytes. Any that are not set (for example
the unused bytes in the address union) are then going to be
uninitialised and contributing to the hash value meaning that
sessions will not be found when they should be.  This impact of this
is that the full lookup would have to be done instead as it was not
found in the flow cache.

VRVDR-52631