npf: Added support for 'auto-per-action' counter type for ACL rules.
All ACL counters (numbered and named) are now maintained in a list in
pmf_group_ext. Named counters can have multiple rules associated with them.
When a group has the auto-per-action option, an action counter (named) is
created when the first rule of that action is created and all subsequent
rules will carry a reference to the same counter.

VRVDR-51137