npf: check IPv4/v6 proto before accessing cached protocol
If TCP flags are configured, it skips adding the separate ncode for
matching on the TCP protocol, as it does a check in npf_match_tcpfl().
This means it will skip a check for the packet being IPv4/IPv6 (i.e. a
non-IP packet), and so add the check into npf_match_tcpfl().

VRVDR-52833