npf: update npf_cache_all() to not convert failures to success
Updated npf_cache_all() to fail with -NPF_RC_NON_IP if handling a non-IP
packet, and with other error codes for header processing failures. When
getting called from npf_hook_notrack(), which is allowed to process
non-IP packets, check for the -NPF_RC_NON_IP error and allow the packet
to be processed by the firewall. Other paths treat it as a failure,
as they expect IP packets.

VRVDR-52833