npf: init cache IP addresses pointer to NULL for embedded pkts
When the embedded packet cache area is initialised, also set the
pointer to the source and destination IP addresses to be NULL. This
is to ensure the unit test in vrvdr-52887 works consistently. There is
not a performance issue as this is only on the path of processing ICMP
error messages.

VRVDR-52833