ut: fix mem leak in the qos class tests
The test that checks that a second system default is not
allowed was configuring an extra map in the test fal after
the setting of the sysdef failed.  Remove the extra line
of config after the failure so that we don't send an unwanted
command that causes us to create unwanted state in the fal that
we don't tidy up.

VRVDR-52631