ut: add mpls test for sending existing route update
The mpls tests did not have a test for sending a route that was
exactly the same as the previous one.  Modify one of the existing
tests to do that by resending the same 'add route command' The
second send needs to do it without verification as that will check
the route does not exist before sending the update.

VRVDR-52791