mpls: track pd state when updating label table entry update
When we update a label_table_entry, we delete the old entry and create
a new one. Copy the pd_state from the old_entry to the new one so that
we can keep the hw entry stats correct.

Before this change the update was decrementing the stats for the value
in the new lable, but that was uninitialised mem, so in some cases it
updated the wrong stat, and in some cases it updated neighbouring memory
causing corruption.

VRVDR-52791