npf: fix issue of crashing when doing firewall logging
A crash can occur during logging when decoding an invalid header within
an ICMP error message. This updates the cache code to return a failure
if it fails to fetch an IP header from the inner packet of an ICMP error
message, so that the logging and other code does not access the unset
cache fields.

VRVDR-52404