commands: remove the code to handle out of order cfg for poe
The controller is now making sure that config for an interface is
delayed until the interface create has been sent. This means that the
features do not need to deal with the feature config arriving before
the interface it is applied to. Remove the code that was doing this
for the interface poe cfg.

Mark the older versions of poe that do not send the interface properly
to the controller as ones that are broken when using this version of
the dataplane.

VRVDR-53049