backplane: add comment about why command replay is needed
The backplane cmd handler has code for checking that the interfaces the
commands use already exist. If they don't then it sets up replay handling
for when they are created so that the command can be applied then.

There is infra in the controller to guarantee that the commands applied to
an interface are sent after the interface has been created, but this only
applies to commands that reference a single interface.

The command handling code only adds the non backplane interfaces to the list
of missing interfaces. The backplane code also registers the event ops,
so that when a backplane interface arrives it can walk all the missing
non backplane interfaces to see if it can now process any of those.

As this command refers to 2 interfaces then the replay code is still needed
so add a comment to make this clear.

VRVDR-53049