rt_commands: remove the code to handle out of order cfg for garp
The controller is now making sure that config for an interface is
delayed until the interface create has been sent. This means that the
features do not need to deal with the feature config arriving before
the interface it is applied to. Remove the code that was doing this
for the interface garp cfg.

VRVDR-53049