vhost: remove the code to handle out of order cfg for vhost
The controller is now making sure that config for an interface is
delayed until the interface create has been sent. This means that the
features do not need to deal with the feature config arriving before
the interface it is applied to.

As the vhost transport link command refers to 2 interfaces there is no
guaraneee from the controller that both will exist.  There is however
a guarantee that the vhost interface will exist, but the transport link
one may not.

The transport link code already has code to deal with link state changes
and that will be triggered when the new interface (transport link) is
created and modified, and therefore we already cope with the out of order
arrival of that.

Remove the code that was caching the config when the vhost was not there
as that is now guaranteed by the controller.

VRVDR-53049