fal: add api to allow dumping of L2 port state
It is possible to debug forwarding issues at L2/L3 by using the vendor
CLI, but this is not very easy to use as it can be hard to track
relationships between different objects due to different names/IDs being
used and different representations of data (hex vs decimal).

Add an API to allow the dumping of a FAL l2 port so that the state can
be displayed via the modelled commands to aid in triage and debugging.
Display this new port state in the ifconfig output.

VRVDR-46947