ptp: select best route for peer
The PTP implementation does not handle multiple masters or asymmetric
routing of PTP traffic.  If a peer has been added on multiple ports, we
need to make an attempt to determine which is "better" based on the FIB.
If the peers are on the backhaul interfaces, then need to be certain
that we select the current active backhaul interface for PTP signaling.

Bugfix: VRVDR-53302
(cherry picked from commit 771fe4cdc7a3a153317916eb2d7df246a1cc8baf)