fal acl: use rule number attr instead of priority in rules
Change the code to set a rule number instead of priority when configuring
a rule in the fal. This makes it easier to debug due to it matching
the configuration, and leaves the hardware-specific code to convert
it into a priority when configuring the hardware.

The attribute FAL_ACL_ENTRY_ATTR_RULE_NUMBER replaces attribute
FAL_ACL_ENTRY_ATTR_PRIORITY.

VRVDR-53381