ut: Verify that ipv4-orig-feat feature is enabled
VRVDR-53369.  'originate' rulesets may be attached to either a dataplane
interface or a loopback interface.  When attached to a dp interface then
check that the feature is enabled only for that interface.   When attached
to a loopback interface ("global:") attach-point then check that the
feature is enabled on all interfaces.