CRYPTO:Introduce batch updates from the server
Introduce a string before each netlink msg indicating either START,
DATA, or END of a batch of policy updates from the xfrm source. The
xfrm source, i.e. strongswan, is responsible for the marshalling of
policies into batches and any relationship between the policies within
a particular batch.

 - START indicates the associated xfrm is the startt of a policy batch
 - DATA inditactes the associated  xfrm is part of a batch
 - END indicates the associated xfrm in the final xfrm in the batch.

Upon completion of processing of the the accompanying netlink policy
if an end of batch was signalled trigger an npf rebuild.

A single policy batch is indicated by an END header.

VRVDR-52437