gpc: add gpc protobuf
Add a protobuf config message for the Generic Packet Classifier which
consists of tables of rules each with matches on various packet header
fields and actions that can be performed on the packet. The initial
functionality covered applies to QoS and ACL but is expected to be expanded
for other features in future.

VRVDR-46091