nh_common: fix hash key during nexthop hash del & add
The nexthop hash key should now include the use of the nexthop list,
but the population of this field was missing in nexthop_hash_del_add
and therefore defaulted to FAL_NHG_USE_IP. So populate use based on
new_nhl.

Reindent the structure initialisation to match dataplane style
guidelines.