route_flags: ignore RTF_UNUSABLE for nexthop comparison
It is essential for PIC-edge operation that nexthops are shared as
much as possible. If the unusable flag is included in the comparison
then since the nexthops created for the RIB update won't include the
flag then if the nexthop for an existing nexthop list isn't usable a
new nexthop list will be created. However, the usability of the
nexthops in that new list will then be evaluated and if not usable the
RTF_UNUSABLE flag will be set.

Further route updates will follow the same process and a new nexthop
list will be created for each causing prefix-dependent convergence and
nexthop scale issues, both in hardware and in the software dataplane
due to sub-optimal algorithms in places.

Therefore, since RTF_UNUSABLE is dataplane-local state ignore it for
the purposes of evalulating whether the nexthop can be shared.

Fixes: VRVDR-53512