ptp: allow ports to not exist (yet)
An interace may not exist when the port, or peer, configuration arrives.
That's not an error since the port is likely to exist at some future point
and the configuration will be replay by the controller when the interface
arrives. While here, ingnore spurious deletes for not yet created ports.

Bugfix: VRVDR-53511