ptp: rework the peer select logic
In this implementation, a peer only has a few states: no_route, routed,
one_hop and switched. one_hop and routed are nominally the same expect
that one is "better" by virtue of having less distance to travel across
the fabric. This change makes the logic in the peer selection clearer.

Bugfix: VRVDR-53538