Dynamic entries in ND cache should not be deleted unless stale
Entries may be deleted if the corresponding entry is removed in the
kernel, even if they are in REACHABLE state. This occurs if the entry
has not been used since the 1 second ageing timer last expired. The
solution is to also check that the entry is STALE before deleting it.

VRVDR-53530