nh_common: ignore path state updates for deleted interfaces
A crash was observed in next_hop_mark_path_state due to derefering the
ifp looked up and with it being NULL. The interface has been deleted
and since the function is being invoked from an interrupt thread it is
expected there could be a window of time between the interface being
deleted and sources updating their datastructures to reflect this.

Address this by checking the ifp and bailing if not found. This could
result in some cases being ignored where the ifindex was never valid,
but we can't easily distinguish between this and the interface having
been deleted so there isn't much we can do.

Fixes: VRVDR-53541