storm_control: don't create duplicate policers when reacting to events
Interface-embellishing features can trigger events notifying that
change to other features, including storm control. The issue is that
such events can be triggered when storm control has already created
FAL objects, and the storm control code was assuming that if it could
be created in the FAL that FAL updates needed to be made, which isn't
true if there are redundant updates.

Therefore, fix this by checking whether a FAL policer object has
already been created for the storm control instance and not issuing a
create in this case. No update is needed since interface-embellishing
feature events don't change any parameters of the policer, only
configuration does.

Fixes: VRVDR-53623