Add support for DEB_BUILD_PROFILES="nocheck"
Debian has support for 'Build Profiles' that allow dependencies and
binary package creation to be made conditional.
See: https://wiki.debian.org/BuildProfileSpec
'nocheck' is a standard one, and if used MUST also be used with
DEB_BUILD_OPTIONS="nocheck".

This will disable the creation of the vyatta-dataplane-test and
vyatta-dataplane-sample-plugin packages, and disable the building
of the associated targets.

Note: that this build profile will not be regularly test built, so may
not be stable. Use at your own risk.