ip_mcast: make use of dp events to improve modularity
Make use of dp_event registrations in the IPv4 and IPv6 multicast code
to receive notifications about init/uninit, interface admin-down and
interface deletion events without calls from the main function and
core interface code.

This improves readability and maintainability of the core interface
code and increases the modularity of the codebase, allowing features
to be more easily added/removed as required.