bridge: remove admin-down ports from VLANs in the FAL
In the case a bridge port is admin-down then we don't expect any
traffic to be received from that port (in any VLAN) and nor do we
expect any traffic to be sent out of that port. However, we program
the VLANs in the hardware regardless, which with a naïve
implementation of hardware, hardware SDK and FAL plugin will result in
the flood set for the VLAN including those admin-down
ports. Admin-down ports being included in the flood set is a
particular problem for hardware where the switch bandwidth is over
subscribed compared to the ports and their maximum speeds, and also it
can result in counters being incremented more than expected.

So to save resources in the hardware programmed by FAL plugins and to
avoid the need for each FAL plugin to have logic to update the flood
sets, then it is desirable to have the interface admin status
influence the FAL programming of the VLAN information for
bridge-ports. To that end, program empty VLAN tagged/untagged sets and
a zero primary-vlan-id when the port is admin-down and react to the
admin-up event by programming the configured VLAN tagged/untagged sets
and primary-vlan-id value.

Fixes: VRVDR-53674