On reboot, disabled bond has members in u/u link state
In CLI when a bonding interface with configured members
is disabled, the members' link state goes to u/D, but
on reboot the members' link state is u/u.

On reboot, the order of adding members and disabling
the bonding interface is reversed. The members are
added to a disabled bonding interface.

In DPDK, the member add API configures the member only
when the bonding device is started. Start the bonding
device if not already started when adding a member.

VRVDR-53373

(cherry picked from commit 7e9e5fa94fd974ee0f5ea287a8fc2834a2f953a2)