ARP dynamic local entries should not be removed due to kernel deletions
Entries in the ARP cache in the dataplane are deleted if the
corresponding entry is removed in the kernel, if the entry was not used
since the 1s periodic timer for ageing last expired. This may impact
performance at high scale due to route churn. A solution is to allow
dynamic entries that were created locally in the dataplane to time out,
unless the deletion is explicitly for local entries as needed by UT.

VRVDR-53536