Minor improvements to lladdr_update() for use by pipelined arp
Prior to the introduction of pipeline arp, lladdr_update() was only
called for newly added entries. As pipeline arp now calls this fn for
existing entries, do not update the expiry time as this is updated in
ll_age() if the entry is in use. Also avoid the debug indicating that
the mac address has moved if it has not changed.

VRVDR-50363