UT:CRYPTO: Handle xfrm direct SA expire messages
The packet and byte expiration limits on an SA are typically
configured as 0 in the crypto unit tests. Unintentionally expire
messages are being generated to the dp_test_xfrm_server from the
dataplane when an SA's stats are updated during forwarding
test. Handle the reception of the expire messages.

VRVDR-45405