if: allow multiple node instance callback to be set
When setting a node instance callback, we use local variablea
to access the array. This will be NULL in the case where
there is a 2nd callback (i.e we have already created the memory
and store it in the gloabl if_node_instance_cleanup_cbs var).
In this case make sure we set the local variables so that we can
properly update.

VRVDR-53593